Add payslip generation panel and functionality
This commit introduces a new panel for generating payslips in both the `Admin` and `User` forms. The `genpayslipPanel` is added to the `FlowLayoutPanel` in the respective designer files. A new event handler, `genpayslip_Click`, is implemented in both `Admin.cs` and `User.cs` to open the payslip generation form. The outdated `govdues_Click` method has been removed from `User.cs`, reflecting the transition to the new payslip generation feature.